requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
PyYAML==6.0.2
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.37.0
pytest==8.3.2
